refactor: tratar avaliacoes

+++ backend/src/main/java/com/denisson/backend/domain/useCases/avaliacoes/DeleteAvaliacaoUseCase.java
@@ -1,5 +1,7 @@
 package com.denisson.backend.domain.useCases.avaliacoes;
 
+import com.denisson.backend.domain.exceptions.NotFoundException;
+import com.denisson.backend.domain.models.Avaliacao;
 import com.denisson.backend.domain.repositories.IAvaliacaoRepository;
 
 public class DeleteAvaliacaoUseCase {
@@ -10,6 +12,12 @@ public class DeleteAvaliacaoUseCase {
     }
 
     public void execute(Long id) {
+        Avaliacao avaliacao = repository.findById(id);
+        
+        if (avaliacao == null) {
+            throw new NotFoundException("Avaliacao não encontrada com id " + id);
+        }
+        
         repository.delete(id);
     }
 }
